% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAERMINUTEQA.R
\name{classifyAERMINUTEDiagnostics}
\alias{classifyAERMINUTEDiagnostics}
\title{Summarize AERMINUTE diagnostic indicators}
\usage{
classifyAERMINUTEDiagnostics(qa)
}
\arguments{
\item{qa}{Output from qaAerminute().}
}
\description{
Summarize AERMINUTE diagnostic indicators
}
\keyword{internal}
