% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAERMINUTEQA.R
\name{classifyAERMINUTEExecution}
\alias{classifyAERMINUTEExecution}
\title{Classify AERMINUTE execution QA}
\usage{
classifyAERMINUTEExecution(qa)
}
\arguments{
\item{qa}{Output from qaAerminute().}
}
\description{
Classify AERMINUTE execution QA
}
\keyword{internal}
