% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAERMINUTEQA.R
\name{classifyAERMINUTEQA}
\alias{classifyAERMINUTEQA}
\title{Classify AERMINUTE QA results}
\usage{
classifyAERMINUTEQA(qa, min_coverage_warn, min_coverage_fail)
}
\arguments{
\item{qa}{Output from qaAerminute().}

\item{min_coverage_warn}{Coverage threshold for warning.}

\item{min_coverage_fail}{Coverage threshold for failure.}
}
\description{
Classify AERMINUTE QA results
}
\keyword{internal}
