% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAERMINUTEQA.R
\name{classifyAERMINUTEReasonableness}
\alias{classifyAERMINUTEReasonableness}
\title{Classify AERMINUTE reasonableness QA}
\usage{
classifyAERMINUTEReasonableness(qa)
}
\arguments{
\item{qa}{Output from qaAerminute().}
}
\description{
Classify AERMINUTE reasonableness QA
}
\keyword{internal}
