% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAERMINUTEQA.R
\name{classifyAERMINUTEStructural}
\alias{classifyAERMINUTEStructural}
\title{Classify AERMINUTE structural QA}
\usage{
classifyAERMINUTEStructural(qa, min_warn, min_fail)
}
\arguments{
\item{qa}{Output from qaAerminute().}

\item{min_warn}{Coverage threshold for warning.}

\item{min_fail}{Coverage threshold for failure.}
}
\description{
Classify AERMINUTE structural QA
}
\keyword{internal}
