% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmet.generics.R
\name{createInput}
\alias{createInput}
\title{Create meteorological model input files}
\usage{
createInput(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
Create meteorological model input files
}
