% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvToISD.R
\name{csvToISD}
\alias{csvToISD}
\title{csvToISD}
\usage{
csvToISD(csv_dfx)
}
\arguments{
\item{csvFile}{is data.frame read in from csv file from NCEI.}
}
\value{
String vector in ISD format.
}
\description{
\code{csvToISD} 

This function takes a data.frame read from csv formatted hourly met data from \url{https://www.ncei.noaa.gov/data/global-hourly/access/}
and converts it to the ISD format detailed in \url{https://www.ncei.noaa.gov/data/global-hourly/doc/isd-format-document.pdf}
}
