% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAERMINUTEQA.R
\name{.assertAERMINUTEArtifacts}
\alias{.assertAERMINUTEArtifacts}
\title{Assert required AERMINUTE artifacts exist}
\usage{
.assertAERMINUTEArtifacts(project_Dir, years)
}
\arguments{
\item{project_Dir}{Base project directory.}

\item{years}{Integer vector of expected years.}
}
\description{
Assert required AERMINUTE artifacts exist
}
\keyword{internal}
