% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFSL.R
\name{downloadFSL}
\alias{downloadFSL}
\title{downloadFSL}
\usage{
downloadFSL(rmetObj, ...)
}
\description{
\code{downloadFSL} DEPRECIATED queries raob database at
the now defunct \url{http://www.esrl.noaa.gov/raobs/intl/GetRaobs.cgi}. 
It then places the resulting text files in subdirectories by year as .FSL files.
}
\seealso{
See \code{downloadIGRA} for upper air path data download.
}
