% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIGRA.R
\name{downloadIGRA}
\alias{downloadIGRA}
\title{downloadIGRA}
\usage{
downloadIGRA(
  rmetObj,
  igraLoc =
    "https://www.ncei.noaa.gov/data/integrated-global-radiosonde-archive/access/data-por/"
)
}
\arguments{
\item{rmetObj}{is valid rmetObject created by createMetProject}

\item{igraLoc}{is the https location of the igra files. By default location is https://www.ncei.noaa.gov/data/integrated-global-radiosonde-archive/access/data-por/}
}
\description{
\code{downloadIGRA} 

The function takes Integrated Global Radiosode Archieve (igra) zip file and extracts the zipfile and
trims the file to match the start date and end date of the 
study period. It adds the local file locations to the rmetObj and returns it.
}
