% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golderPlot.R
\name{golderPlot}
\alias{golderPlot}
\title{golderPlot}
\usage{
golderPlot(surface_file_1)
}
\description{
\code{golderPlot} takes data frame from \code{surfReader} produces plot of 
surface roughness length vs. 1/Monin Obukhov Length
}
