% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmet.generics.R
\name{locYears}
\alias{locYears}
\title{List years covered by an rmet project}
\usage{
locYears(rmetObj)
}
\arguments{
\item{rmetObj}{An \code{rmet} object}
}
\value{
Character vector of years
}
\description{
List years covered by an rmet project
}
