% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFinal.R
\name{makeFinal.rmet}
\alias{makeFinal.rmet}
\title{makeFinal}
\usage{
makeFinal.rmet(rmetObj, outfilePrefix = "outFile")
}
\description{
\code{makeFinal} reads the surface file output of aermet
 across multiple years
}
