% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapstation.R
\name{mapstation}
\alias{mapstation}
\title{Map station around a point of interest}
\usage{
mapstation(site, farthest = 25, engine = "leaflet", labels = "popup")
}
\arguments{
\item{site}{a character string that can be geocoded by the Google Maps API.
For example, "30341", "4770 Buford Hwy. Atlanta GA 30341", and "CDC".}

\item{farthest}{a numeric value in meters that determine the farthest station
to be mapped from the \code{site} If no value is specified a defalt value
of 25 kilometers is used.}

\item{engine}{a character string to specify the mapping application to be used:
\code{"leaflet"}.}
}
\description{
\code{mapstation} Creates a basic map with the stations around a site of interest.
The map shows the site of interest in red and the stations in blue popups. Popups
can be clicked for additional information about each particular station. Station
information is obtained from \url{http://www1.ncdc.noaa.gov/pub/data/noaa/isd-history.txt})
}
