% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmet.generics.R
\name{print.rmet}
\alias{print.rmet}
\title{Print an rmet object}
\usage{
\method{print}{rmet}(x, ...)
}
\arguments{
\item{x}{An \code{rmet} object}

\item{...}{Unused}
}
\description{
Print an rmet object
}
