% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMet.R
\name{processMet}
\alias{processMet}
\title{processMet}
\usage{
processMet(
  rmetObj,
  processor = c("aerminute", "aersurface", "aermet1", "aermet2", "aermet3"),
  newaersurface = TRUE
)
}
\description{
\code{processMet} This executes the AERMINUTE, AERSUFACE or AERMET prcoessors
using appropriately initialized \code{rmet} object. This means that inputFiles 
have been created (typically using \code{createInput}) and written (using \code{writeInput}).
}
