% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaAerminute.R
\name{qaAerminute}
\alias{qaAerminute}
\title{qaAerminute}
\usage{
qaAerminute(rmetObj)
}
\arguments{
\item{rmetObj}{a rmet class object}
}
\description{
\code{qaAerminute} reads QA Information from Aerminute Run This function returns a list object containing run log of aerminute as tables.
\itemize{
\item{\code{RecCheck} Record format QA Output, }
\item{\code{Flags} QA flag summary,}
\item{\code{qaCalms} Standard calms summary,}
\item{\code{qaMissing} Missing winds summary,}
\item{\code{qaVariable} Variable winds summary,}
\item{\code{qaValid} Standard valid winds summary,}
\item{\code{qaVarWS} Comparison of minute to hourly surface data - wind speed (variable winds),}
\item{\code{qaValWD} Comparison of minute to hourly surface data - wind speed (valid winds),}
\item{\code{qaValWD} Comparison of minute to hourly surface data - wind direction (valid winds),}
\item{\code{YearNumSum} Summary of available hours for model,}
\item{\code{monthlyData} Availability by month for all years.}
}
}
