% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMergeTrim.R
\name{rasterMergeTrim}
\alias{rasterMergeTrim}
\title{rasterMergeTrim}
\usage{
rasterMergeTrim(rasterFiles, long, lat, buffer = 5000, outFile)
}
\description{
\code{rasterMergeTrim} Crops a NLCD or NED file to a radius around a fixed point.
The point is expected to be WGS 1984 lattitude and longitude. Buffer is in meters.

NLCD 1992 data is becoming challenging to obtain, at least for me. Previously,
I have been downloading NLCD 1992 data from Earth Explorer. These files have 
colortable attributes and will it will be preserved in the trimming and merging
process. However, it seems these files have wandered off of the Earth Explorer website,
so not, I am currently downloading these rasters from [Mult-Resolution Land Characteristics Consortium](http://www.mrlc.gov/nlcd92_data.php)
However, these tiles lack a legend and colortable attributes. This function will add the
a colortable for processing with AERSURFACE. 

 @param rasterFiles a single character string or vector of character strings specifying 
 the location of the rasters to be cropped merged. 
 
 @param long numeric longitude of point
 @param lat numeric latitude of point
 @buffer distance in meters for cropping around point (meters)
 @outFile destination of merged and cropped output
}
