% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRmet.R
\name{saveRmet}
\alias{saveRmet}
\title{Save an rmet project to disk}
\usage{
saveRmet(rmetObj)
}
\arguments{
\item{rmetObj}{An rmet object}
}
\value{
Invisibly returns the rmet object
}
\description{
Save an rmet project to disk
}
