% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfCheck.R
\name{surfCheck}
\alias{surfCheck}
\title{surfCheck}
\usage{
surfCheck(
  rmetObj,
  outfile = "surfaceFile.pdf",
  file.choose = FALSE,
  returnDfx = FALSE
)
}
\description{
\code{surfCheck} Uses \code{surfReader} to read surface files then plot
}
