% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAERMINUTEQA.R
\name{writeAERMINUTEQAReport}
\alias{writeAERMINUTEQAReport}
\title{Write AERMINUTE QA report}
\usage{
writeAERMINUTEQAReport(
  rmetObj,
  file = file.path(paste(rmetObj$project_Dir, "output", "QA_AERMINUTE.txt", sep = "/"))
)
}
\arguments{
\item{rmetObj}{An rmet object with classified AERMINUTE QA results.}

\item{file}{Output file path.}
}
\description{
Write AERMINUTE QA report
}
\keyword{internal}
