% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmet.generics.R
\name{writeInputFile}
\alias{writeInputFile}
\title{Write model input files to disk}
\usage{
writeInputFile(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
Write model input files to disk
}
