% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInputFile.R
\name{writeInputFile.rmet}
\alias{writeInputFile.rmet}
\title{writeInputFile}
\usage{
\method{writeInputFile}{rmet}(rmetObj, type = c("aerminute", "aersurface"), dewinter = FALSE)
}
\description{
\code{writeInputFile}
}
