% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInput.R
\name{createInput.rmet}
\alias{createInput.rmet}
\title{createInput}
\usage{
\method{createInput}{rmet}(
  rmetObj,
  type = c("aerminute", "aersurface_nws", "aermet1", "aermet2", "aermet3"),
  ustar = TRUE,
  inputOps = list(S1 = list(), S2 = list(), S3 = list(other = list("THRESH_1MIN 0.5"),
    method = list("REFLEVEL SUBNWS", "STABLEBL ADJ_U*")))
)
}
\description{
\code{createInput}
}
