% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFSL.R
\name{downloadFSL}
\alias{downloadFSL}
\title{downloadFSL}
\usage{
downloadFSL(rmetObj, ...)
}
\description{
\code{downloadFSL} querries raob database at
\url{http://www.esrl.noaa.gov/raobs/intl/GetRaobs.cgi}. 
It then places the resulting text files in subdirectories by year as .FSL files.
}
