% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIGRA.R
\name{downloadIGRA}
\alias{downloadIGRA}
\title{downloadIGRA}
\usage{
downloadIGRA(
  rmetObj,
  igraLoc =
    "https://www.ncei.noaa.gov/data/integrated-global-radiosonde-archive/access/data-por/"
)
}
\description{
Downloads, extracts, and trims Integrated Global Radiosonde Archive (IGRA)
upper-air data to match the rmet project date range.
}
