% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTD3505.R
\name{downloadTD3505}
\alias{downloadTD3505}
\title{downloadTD3505}
\usage{
downloadTD3505(rmetObj, check = TRUE, ...)
}
\description{
\code{downloadTD3505} Boo!

The function is designed to perform basic checks to make sure that 
the data are valid (e.g. WBAN and USAF id's match \url{http://www1.ncdc.noaa.gov/pub/data/noaa/isd-history.txt})
}
