% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTD6401.R
\name{downloadTD6401}
\alias{downloadTD6401}
\title{downloadTD6401}
\usage{
downloadTD6401(rmetObj, check = TRUE, ...)
}
\description{
\code{downloadTD6401} 

The function is designed to perform basic checks to make sure that 
the data are valid (e.g. WBAN and USAF id's match \url{http://www1.ncdc.noaa.gov/pub/data/noaa/isd-history.txt})
}
