% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTD6405.R
\name{downloadTD6405}
\alias{downloadTD6405}
\title{downloadTD6405}
\usage{
downloadTD6405(rmetObj, check = TRUE, ...)
}
\arguments{
\item{rmetObj}{A valid rmet object}

\item{check}{Logical; if TRUE (default) only missing/invalid files are downloaded}

\item{...}{Unused}
}
\value{
Updated rmet object
}
\description{
Downloads NOAA TD6405 ASOS 1-minute surface data and validates file integrity.
}
