% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFinal.R
\name{makeFinal.rmet}
\alias{makeFinal.rmet}
\title{makeFinal}
\usage{
makeFinal.rmet(rmetObj, outfilePrefix = "outFile")
}
\arguments{
\item{rmetObj}{{An object of class \code{rmet}.}}

\item{outFilePrefix}{is the name of the profile and surface file. For
instance, \dQuote{KFTY} will result in \dQuote{KFTY.sfc} and \dQuote{KFTY.pfl}.}
}
\description{
\code{makeFinal} reads the surface file and profile output of aermet
 across multiple years and creates a multi-year surface files. This function
 works by scanning the project directory year files for .sfc and .pfl file
 output and appending them together into single files and writing them at the
 rmetObj project directory.
}
