% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesoData.r
\name{mesoData}
\alias{mesoData}
\title{mesoData}
\usage{
mesoData(station = NULL, rmetObj = NULL, API_KEY, type = c("surf"))
}
\arguments{
\item{station}{a 4 letter ICAO character string.
For example, "KATL" for Altanta.}

\item{rmetObj}{is an rmet class list created by createMetProject (optional).}
}
\description{
This function will query the mesowest API and extract location and elevation (in meters) 
of surface station using the  International Civil Aviation Organization (ICAO) 
code. If a rmet class object is passed as well, it will append the information to the rmet list and return it.

\code{mesoData}
}
