% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMet.R
\name{processMet}
\alias{processMet}
\title{processMet}
\usage{
processMet(
  rmetObj,
  processor = c("aerminute", "aersurface", "aermet1", "aermet2", "aermet3"),
  newaersurface = TRUE
)
}
\description{
\code{processMet} Boo!
}
